BYTE instead of int.

+++ Code/RelayDemo.c
@@ -6,8 +6,6 @@
 #include "Include/TCPIPConfig.h"
 #include "Include/TCPIP_Stack/TCPIP.h"
 
-
-
 #if defined(STACK_USE_DHCP_RELAY)
 
 #include "Include/TCPIP_Stack/TCPIP.h"
@@ -32,7 +30,7 @@ static int Arp();
 static void RelayToServer(BOOTP_HEADER *Header, int type);
 static void RelayToClient(BOOTP_HEADER *Header, int type);
 void Log(char *top, char *bottom);
-void LogMac(int pos, BOOTP_HEADER *Header);
+void LogMac(BYTE pos, BOOTP_HEADER *Header);
 
 void DHCPRelayTask(void)
 {
@@ -358,7 +356,7 @@ void Log(char *top, char *bottom){
  * @param pos    Start position
  * @param Header Header containing the MAC-address
  */
-void LogMac(int pos, BOOTP_HEADER *Header){
+void LogMac(BYTE pos, BOOTP_HEADER *Header){
 	DisplayWORD(pos, Header->ClientMAC.v[4]);
 	DisplayWORD(pos+4, Header->ClientMAC.v[5]);
 }
